requirements.txt
selenium
webdriver-manager
pandas
beautifulsoup4
requests
sqlalchemy
mysql-connector-python
python-dotenv